Add separate controller for OmniAuth
